animated path: Allow to disable animations for the layer
Add a prop to disable animation on the `AnimatedArrowPathLayer`. This
deactivates the animated frame request and arrow path time step, thus
saving the GPU from calculations. The previous speed divider of 0 also
had the effect of making the path not move, but the layer was still
refreshed for every frame, impacting GPU performances.