deck.gl: Remove unused layers from config file
`isochronePolygons`, `transitStations` and `transitStationSelected`
layers are not used anymore, so they are removed from the layers
configuration file, as well as from the default preferences.

Also, the path stroke layers have been removed and, if necessary, can be
replaced with some of deck.gl's feature directly in the main
corresponding layer.