preferences: convert rgba color preferences to hex
The migration to deck.gl changed the format of the colors to the more
universal hexadecimal rgba representation (starting with #). Previous
color preferences were in the format of a rgba function (for example
`rgba(255,0,0,0.5)`). This old format is automatically converted to hex
string when we get a preferences that has the name color and the rgba
format. This commit can be reverted after a while, when all current
deck.gl users have changed their preferences.